Move test helpers to separate folder
